address review comments
- move `check_transaction_validity_period` to the utils instead of making it a member function of `ChainStoreAdapter`
- revert moving the `get_block_header_on_chain_by_height` and associated changes
